requirements.txt
numpy
pandas
matplotlib

iPython
jupyter
tqdm
requests

tensorflow
librosa
PyYAML
xlwt

autopep8